fix bugs

+++ scr/sml.c
@@ -17,7 +17,6 @@ void submit_vote(Protected *p) {
 void create_block(CellTree *tree, Key *author, int d) {
     Block *block = (Block *) malloc(sizeof(Block));
     CellProtected *lcp = read_protected("Pending_votes.txt");
-    CellTree *lastNode = last_node(tree);
     block->author = author;
     block->nonce = 0;
     block->votes = lcp;
@@ -27,17 +26,24 @@ void create_block(CellTree *tree, Key *author, int d) {
     system("rm Pending_votes.txt");
     compute_proof_of_work(block, d);
     write_block("Pending_block", block);
-    add_child(lastNode, create_node(block));
+    CellTree *new = create_node(block);
+    if (tree->block) {
+        CellTree *lastNode = last_node(tree);
+        add_child(lastNode, new);
+    } else {
+        tree->block = block;
+    }
 }
 
 void add_block(int d, char *name) {
     char path[1 << 16];
-    strcpy(path, "./Blockchain/");
-    strcat(path, name);
+    sprintf(path, "./Blockchain/%s", name);
     FILE *fWrite = fopen(path, "w");
     FILE *fRead = fopen("Pending_block", "r");
+    Block *block = read_block("Pending_block");
+    if ((!block) || (!verify_block(block, 4))) exit(5);
     char buf[1 << 16];
-    while (fgets(buf, 1 << 16, fRead) > 0) {
+    while (fgets(buf, 1 << 16, fRead)) {
         fputs(buf, fWrite);
     }
     system("rm Pending_block");
@@ -45,8 +51,7 @@ void add_block(int d, char *name) {
     fclose(fRead);
 }
 
-
-int strcmp_unsigned(const unsigned char *s1, const unsigned  char *s2) {
+int strcmp_unsigned(const unsigned char *s1, const unsigned char *s2) {
     unsigned char *p1 = (unsigned char *) s1;
     unsigned char *p2 = (unsigned char *) s2;
     while ((*p1) && (*p1 == *p2)) {
@@ -55,6 +60,7 @@ int strcmp_unsigned(const unsigned char *s1, const unsigned  char *s2) {
     }
     return (*p1 - *p2);
 }
+
 // LECTURE DE L’ARBRE ET CALCUL DU GAGNANT
 CellTree *read_tree() {
     DIR *rep = opendir("./Blockchain/");
@@ -92,7 +98,6 @@ CellTree *read_tree() {
     }
     return NULL;
 }
-
 
 Key *compute_winner_BT(CellTree *tree, CellKey *candidates, CellKey *voters, int sizeC, int sizeV) {
     CellProtected *lcp = fusionAll(tree);
